pyproject.toml
[project]
name = "ocr-pipeline"
version = "1.0.0"
description = "OCR Pipeline für Auftragsbestätigungen (Gemini/Mistral)"
requires-python = ">=3.10"
dependencies = [
    "streamlit>=1.52.0",
    "python-dotenv>=1.2.0",
    "pydantic>=2.0.0",
    "PyMuPDF>=1.26.0",
    "Jinja2>=3.1.0",
    "openai>=1.0.0",
    "google-genai>=0.3.0",
    "mistralai>=0.1.0",
    "google-cloud-aiplatform>=1.38.0"
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-cov>=4.0.0",
]

[project.scripts]
batch-runner = "batch_runner:main"

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
log_cli = true
